requirements.txt
altair==5.5.0
anyio==4.7.0
asgiref==3.8.1
attrs==24.2.0
blinker==1.9.0
boto3==1.36.22
botocore==1.36.22
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.2
Django==5.1.4
django-cors-headers==4.6.0
django-extensions==3.2.3
django-redis==5.4.0
djangorestframework==3.15.2
et_xmlfile==2.0.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwt==1.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mypy==1.15.0
mypy-extensions==1.0.0
mysqlclient==2.2.6
narwhals==1.15.1
numpy==2.1.3
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pip==24.2
plotly==5.24.1
protobuf==5.29.0
pyarrow==18.1.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.1
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.0
s3transfer==0.11.2
scikit-learn==1.6.0
scipy==1.14.1
selenium==4.27.1
setuptools==75.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.2
streamlit==1.40.2
tenacity==9.0.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.2
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==6.0.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wheel==0.44.0
wsproto==1.2.0
